+++ examples/python_example.py
@@ -7,10 +7,12 @@ def print_stats(mine):
     print "MEV", mine.mev()
     print "MCN (eps=0)", mine.mcn(0)
     print "MCN (eps=1-MIC)", mine.mcn_general()
+    print "GMIC", mine.gmic()
+    print "TIC", mine.tic()
 
 x = np.linspace(0, 1, 1000)
 y = np.sin(10 * np.pi * x) + x
-mine = MINE(alpha=0.6, c=15)
+mine = MINE(alpha=0.6, c=15, est="mic_approx")
 mine.compute_score(x, y)
 
 print "Without noise:"