+++ examples/relationships.py
@@ -4,11 +4,13 @@ import matplotlib.pyplot as plt
 from minepy import MINE
 
 
+rs = np.random.RandomState(seed=0)
+
 def mysubplot(x, y, numRows, numCols, plotNum,
               xlim=(-4, 4), ylim=(-4, 4)):
 
     r = np.around(np.corrcoef(x, y)[0, 1], 1)
-    mine = MINE(alpha=0.6, c=15)
+    mine = MINE(alpha=0.6, c=15, est="mic_approx")
     mine.compute_score(x, y)
     mic = np.around(mine.mic(), 1)
     ax = plt.subplot(numRows, numCols, plotNum,
@@ -30,24 +32,24 @@ def mvnormal(n=1000):
     cors = [1.0, 0.8, 0.4, 0.0, -0.4, -0.8, -1.0]
     for i, cor in enumerate(cors):
         cov = [[1, cor],[cor, 1]]
-        xy = np.random.multivariate_normal([0, 0], cov, n)
+        xy = rs.multivariate_normal([0, 0], cov, n)
         mysubplot(xy[:, 0], xy[:, 1], 3, 7, i+1)
 
 def rotnormal(n=1000):
     ts = [0, np.pi/12, np.pi/6, np.pi/4, np.pi/2-np.pi/6,
           np.pi/2-np.pi/12, np.pi/2]
     cov = [[1, 1],[1, 1]]
-    xy = np.random.multivariate_normal([0, 0], cov, n)
+    xy = rs.multivariate_normal([0, 0], cov, n)
     for i, t in enumerate(ts):
         xy_r = rotation(xy, t)
         mysubplot(xy_r[:, 0], xy_r[:, 1], 3, 7, i+8)
 
 def others(n=1000):
-    x = np.random.uniform(-1, 1, n)
-    y = 4*(x**2-0.5)**2 + np.random.uniform(-1, 1, n)/3
+    x = rs.uniform(-1, 1, n)
+    y = 4*(x**2-0.5)**2 + rs.uniform(-1, 1, n)/3
     mysubplot(x, y, 3, 7, 15, (-1, 1), (-1/3, 1+1/3))
     
-    y = np.random.uniform(-1, 1, n)
+    y = rs.uniform(-1, 1, n)
     xy = np.concatenate((x.reshape(-1, 1), y.reshape(-1, 1)), axis=1)
     xy = rotation(xy, -np.pi/8)
     lim = np.sqrt(2+np.sqrt(2)) / np.sqrt(2)
@@ -57,15 +59,15 @@ def others(n=1000):
     lim = np.sqrt(2)
     mysubplot(xy[:, 0], xy[:, 1], 3, 7, 17, (-lim, lim), (-lim, lim))
     
-    y = 2*x**2 + np.random.uniform(-1, 1, n)
+    y = 2*x**2 + rs.uniform(-1, 1, n)
     mysubplot(x, y, 3, 7, 18, (-1, 1), (-1, 3))
     
-    y = (x**2 + np.random.uniform(0, 0.5, n)) * \
-        np.array([-1, 1])[np.random.random_integers(0, 1, size=n)]
+    y = (x**2 + rs.uniform(0, 0.5, n)) * \
+        np.array([-1, 1])[rs.random_integers(0, 1, size=n)]
     mysubplot(x, y, 3, 7, 19, (-1.5, 1.5), (-1.5, 1.5))
 
-    y = np.cos(x * np.pi) + np.random.uniform(0, 1/8, n)
-    x = np.sin(x * np.pi) + np.random.uniform(0, 1/8, n)
+    y = np.cos(x * np.pi) + rs.uniform(0, 1/8, n)
+    x = np.sin(x * np.pi) + rs.uniform(0, 1/8, n)
     mysubplot(x, y, 3, 7, 20, (-1.5, 1.5), (-1.5, 1.5))
 
     xy1 = np.random.multivariate_normal([3, 3], [[1, 0], [0, 1]], int(n/4))